Updated theme for next release of magit
This updates the Solarized theme for the 'next' branch of magit. I
prioritized adhering to magit's current color usage, overriding with the
respective Solarized colors. Second priority was keeping the look of the
current version of magit, especially boxes around references in the
log. All magit faces which defined a background/foreground are defined
here with the respective Solarized adaptive color, and the vast majority
of faces that inherit from others are not overridden.

Since the 'next' branch has not yet been released, this PR should not be
merged until that time, as it removes the support for the current
version of magit to avoid conflicts. However, if someone wants to take
the time to do the comparison, it should be possible to keep current
support for all non-conflicting faces.