use proper section headings to separate packages
Instead of e.g.

     ;; package

use

;;;;; package

This allows collapsing sections similar to how this is done in Org
mode.  The outshine minor-mode can be used to achieve them same in
Emacs-Lisp and other major-modes.

The "Theme Faces" section has to subsections "Built-in" and
"Third-party".  Unfortunately the latter contains many subsubsections
for sections that are actually part of Emacs.  That should be fixed
over time.