Do not depend on the f and s packages for tests
`s' was not used directly but is a dependency of `f'.  The only
function from `f' which was used here is `f-expand' which is
defined like so:

  (defun f-expand (path &optional dir)
    "Expand PATH relative to DIR (or `default-directory')."
    (directory-file-name (expand-file-name path dir)))

Since the `directory-file-name' part is irrelevant in our usage,
we might as well just use `expand-file-name'.

Also remove some empty lines and the empty and not properly
terminated "Commentary" section.