Autoload solarized-color-blend to avoid breaking customisation
When saving customisation after enabling this theme, the saved settings for certain variables (e.g. the highlight symbol faces) get set to forms which reference solarized-color-blend. If the saved theme is later switched to a different theme, then those variables settings break, because solarized will not be loaded at the time they are evaluated.