Collect all basic faces in a single section and sort that
Collect all faces whose Custom group is `basic-faces' in the very
first section that contains face definitions.

Actually there is one except - `isearch' is placed in the "isearch"
section because another Isearch face is being themed by this theme
and that face (like all other Isearch faces except `isearch') are
only members of the `isearch' group, but not of `basic-faces'.  So
this inconsistency in Solarized is due to an inconsistency in Emacs.

Note that Solarized does not theme the following members of the
`basic-faces' group:
- `bold'
- `bold-italic'
- `border'
- `fixed-pitch'
- `fixed-pitch-serif'
- `glyphless-char'
- `italic'
- `mode-line-emphasis'
- `mode-line-highlight'
- `mouse'
- `nobreak-space'
- `scroll-bar'
- `tool-bar'
- `tool-tips'
- `tty-menu-enabled'
- `tty-menu-selected'
- `underline'
- `variable-pitch'
- `window-divider'
- `window-divider-first-pixel'
- `window-divider-last-pixel'