Adjust magit-blame faces to recent upstream changes
The hash/name/date/summary faces no longer inherit from
`magit-blame-heading'.  Instead they are combined with that or another
face like `:face (list BG-FACE %-FACE)' or in some cases they are not
combined with another face.

The heading face should only set the background and the
heading and the new `magit-blame-margin' should inherit from that.

It isn't necessary to theme that or the also new `magit-blame-dimmed'
here because these faces (and `magit-blame-heading') now `:weight
normal :normal' by default.