Do not derive outline faces from org faces
Org is based on Outline -- not the other way around, so we cannot
assume that the Org faces are available when someone is using
outline-mode.

This can be fixed either by inverting the direction of the :inherit
or by theming the two groups of faces independently.  This commit
does the later.

Additionally this adds a new option `solarized-scale-outline-headlines',
which is like the existing `solarized-scale-org-headlines', just for
another package.

Fixes #308.